Weekly Calendar Formatting in ProfilePage
